Paging

+++ src/main/com/excilys/controller/CDBController.java
@@ -1,10 +1,10 @@
 package main.com.excilys.controller;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import main.com.excilys.model.Company;
 import main.com.excilys.model.Computer;
+import main.com.excilys.model.Page;
 import main.com.excilys.service.CompanyService;
 import main.com.excilys.service.ComputerService;
 import main.com.excilys.view.CDBView;
@@ -14,11 +14,19 @@ public class CDBController {
 	CDBView view;
 	CompanyService cpnService;
 	ComputerService cptService;
+	Page page;
+	
+	private int currentPage;
+
+	
+	private final int LIMIT=10;
+	private final int CURRENT_PAGE=1;
 	
 	public CDBController(CDBView view, CompanyService cpnService, ComputerService cptService){
 		this.view=view;
 		this.cpnService=cpnService;
 		this.cptService=cptService;
+		page = new Page(LIMIT,CURRENT_PAGE);
 	}
 	
 	/**
@@ -62,18 +70,31 @@ public class CDBController {
 	 * option 1
 	 */
 	public void listComputers() {
-		List<Computer> computers = new ArrayList<Computer>();
-		computers = this.cptService.getComputers();
-		this.view.displayComputers(computers);
+		currentPage =1;
+		boolean ok=true;
+		List<Computer> computers = null;
+		while(ok) {
+			computers = this.cptService.getComputers(LIMIT,currentPage);
+			if(computers.isEmpty())	ok=false;
+			page.setCurrentPage(currentPage++);
+			this.view.displayComputers(computers,page);
+		}
+		
 	}
 	
 	/**
 	 * option 2
 	 */
 	public void listCompanies() {
-		List<Company> companies = new ArrayList<Company>();
-		companies = this.cpnService.getCompanies();
-		this.view.displayCompanies(companies);
+		currentPage =1;
+		boolean ok=true;
+		List<Company> companies = null;
+		while(ok) {
+			companies = this.cpnService.getCompanies(LIMIT, currentPage);
+			if(companies.isEmpty()) ok=false;
+			page.setCurrentPage(currentPage++);
+			this.view.displayCompanies(companies,page);
+		}
 	}
 	
 	/** 
@@ -83,9 +104,7 @@ public class CDBController {
 		Computer computer = null;
 		Long id = this.view.queryId();
 		if(id!=null) computer = this.cptService.findById(id);
-		if(computer!=null) {
-			this.view.displayComputer(computer);
-		}
+		if(computer!=null)	this.view.displayComputer(computer);
 	}
 	
 	/**
@@ -93,7 +112,7 @@ public class CDBController {
 	 */
 	public void createComputer() {
 		Computer computer =null;
-		computer = this.view.queryComputerToCreate();
+		computer = this.queryComputerToCreate();
 		if(computer!=null) this.cptService.createComputer(computer);
 	}
 	
@@ -102,16 +121,64 @@ public class CDBController {
 	 * option 5
 	 */
 	public void updateComputer() {
-		Computer computer = this.view.queryComputerToUpdate();
-		this.cptService.update(computer);	
+		Computer computer = null;
+		computer = this.queryComputerToUpdate();
+		if(computer!=null) this.cptService.update(computer);	
 	}
 	
 	/**
 	 * option 6
 	 */
 	public void deleteComputer() {
-		Long id = this.view.queryComputerToDelete();
-		this.cptService.delete(id);	
+		Long id = null;
+		id = this.queryComputerToDelete();
+		if(id!=null) this.cptService.delete(id);	
+	}
+	
+	
+	/**
+	 * Asks for the computer information
+	 * @return a Computer object
+	 */
+	public Computer queryComputerToCreate() {
+		Computer computer = null;
+		String name =view.queryName();
+		if(name!=null && !name.isEmpty()) {
+			computer = new Computer();
+			computer.setName(name);
+			computer.setIntroduced(view.queryDate());
+			computer.setDiscontinued(view.queryDate());
+			computer.setCompany(new Company());
+			computer.getCompany().setId(view.queryId());
+		} else {
+			view.notifyInvalidName();
+		}
+		return computer;
+	}
+	
+	/**
+	 * Asks for the computer information
+	 * @return a Computer object
+	 */
+	public Computer queryComputerToUpdate() {
+		Computer computer = new Computer();
+		
+		computer.setId(view.queryId());
+		computer.setName(view.queryName());
+		computer.setIntroduced(view.queryDate());
+		computer.setDiscontinued(view.queryDate());
+		computer.getCompany().setId(view.queryId());
+
+		return computer;
+	}
+	
+
+	/**
+	 * Asks for an id
+	 * @return the id
+	 */
+	public Long queryComputerToDelete() {
+		return view.queryId();
 	}
